+++ Wrappers/Python/cil/plugins/tigre/__init__.py
@@ -19,4 +19,5 @@
 from .Geometry import CIL2TIGREGeometry
 from .ProjectionOperator import ProjectionOperator
 from .FBP import FBP
-from .Algorithms import ART, SART, SIRT, OSSART
+from .ARTAlgorithms import ART, SART, SIRT, OSSART
+from .Algorithms import tigre_algo_wrapper
